+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/integrations/ae2/AEFluidKeyFactory.kt
@@ -1,0 +1,9 @@
+package site.siredvin.peripheralworks.integrations.ae2
+
+import appeng.api.stacks.AEFluidKey
+import site.siredvin.broccolium.modules.storage.fluid.AgnosticFluidStack
+import site.siredvin.broccolium.modules.storage.fluid.toVariant
+
+object AEFluidKeyFactory {
+    fun of(stack: AgnosticFluidStack): AEFluidKey = AEFluidKey.of(stack.toVariant())!!
+}