+++ projects/fabric/src/testMod/kotlin/site/siredvin/peripheralworks/testmod/FabricPeripheralWorksTestMod.kt
@@ -17,8 +17,11 @@ object FabricPeripheralWorksTestMod : ModInitializer {
             CctFixtureCommands.importFiles(it)
         }
         Testiarium.register(PeripheralWorksGameTests::class.java)
-        Testiarium.register(AE2ConfigurableObjectsGameTests::class.java)
-        Testiarium.register(AE2WirelessTerminalGameTests::class.java)
+        if (FabricLoader.getInstance().isModLoaded("ae2")) {
+            Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.AE2GameTests"))
+            Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.AE2ConfigurableObjectsGameTests"))
+            Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.AE2WirelessTerminalGameTests"))
+        }
         if (FabricLoader.getInstance().environmentType == EnvType.CLIENT) {
             Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.NetworkManagerClientGameTests"))
         }