+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/ForgePeripheralWorks.kt
@@ -63,11 +63,12 @@ object ForgePeripheralWorks {
 
     init {
         ForgePeripheralium.sayHi()
+        PeripheralWorksCore.configure(ForgeModPlatform, ForgeModRecipeIngredients, ForgeModBlocksReference)
+        loader.maybeLoadIntegration("ae2", "Registration").ifPresent { (it as Runnable).run() }
         // Configure configuration
         @Suppress("DEPRECATION", "removal")
         val context = ModLoadingContext.get()
         context.registerConfig(ModConfig.Type.COMMON, ConfigHolder.commonSpec, "${PeripheralWorksCore.MOD_ID}.toml")
-        PeripheralWorksCore.configure(ForgeModPlatform, ForgeModRecipeIngredients, ForgeModBlocksReference)
         ForgeNetworkHandler.setup()
         val eventBus = MOD_CONTEXT.getKEventBus()
         eventBus.addListener(this::commonSetup)
@@ -93,6 +94,7 @@ object ForgePeripheralWorks {
         loader.maybeLoadIntegration("occultism").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("easy_villagers").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("toms_storage").ifPresent { (it as Runnable).run() }
+        loader.maybeLoadIntegration("ae2").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("mna").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("deepresonance").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("powah").ifPresent { (it as Runnable).run() }
@@ -125,7 +127,6 @@ object ForgePeripheralWorks {
         loader.maybeLoadIntegration("naturescompass").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("ars_nouveau").ifPresent { (it as Runnable).run() }
         loader.maybeLoadIntegration("projecte").ifPresent { (it as Runnable).run() }
-        loader.maybeLoadIntegration("ae2").ifPresent { (it as Runnable).run() }
     }
 
     @Suppress("MemberVisibilityCanBePrivate")