+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/ae2/AE2Setup.kt
@@ -1,0 +1,11 @@
+package site.siredvin.peripheralworks.integrations.ae2
+
+import appeng.api.features.P2PTunnelAttunement
+
+object AE2Setup {
+    fun registerPeripherals() = Unit
+
+    fun registerAttunement() {
+        P2PTunnelAttunement.registerAttunementTag(Registration.WIRED_NETWORK_P2P_TUNNEL.get())
+    }
+}